% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsDownload.R
\name{cdsDownload}
\alias{cdsDownload}
\title{Download of the climate data.}
\usage{
cdsDownload(U_ID, API_Key, lat, lon, sPath)
}
\arguments{
\item{U_ID}{The user ID for the authentication in the Coprnicus CDS service (string)}

\item{API_Key}{The key needed to make an API request (string)}

\item{lat}{The latitude of the site of interest, expressed in decimals with a single decimal value (integer)}

\item{lon}{The longitude of the site of interest, expressed in decimals with a single decimal value (integer)}

\item{sPath}{The path of the directory where the file is going to be saved (string)}
}
\value{

}
\description{
Create a .nc file, containing the mean temperature and precipitation records of a defined location,
downloaded from the ECMFWR "ERA-5 Land monthly averaged data", via the Copernicus CDS API.
Before using, the registration to the Copernicus CDS site is necessary.
}
\examples{
cdsDownload(U_ID = "85930", API_Key = "4017122e-9803-4938-83a1-e2b75f6e2d9b",
                      lat = 44.2, lon = 11.32, sPath = "C:/Users/seba_/Documents/Cestino_documenti")
}
