% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mast.R
\name{mast}
\alias{mast}
\title{Mast probability calculation}
\usage{
mast(filePath)
}
\arguments{
\item{filePath}{The path to the .nc file to be used as input for the mast calculation}
}
\value{
The function returns a table with 2 columns, the first one with the years and the second one with
the associated probability of the mast event
}
\description{
Calculate the mast probability for the current year, using as input the file downloaded from the Copernicus
CDS API. It prints a plot representing a time series of the mast event probabilty from the first available
data to the current year
}
\examples{
mast(filePath = "C:/Users/seba_/Documents/Cestino_documenti/85930_t2s_tp.nc")
}
