% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastFaSyl.R
\name{mastFaSyl}
\alias{mastFaSyl}
\title{Beech mast probability calculation}
\usage{
mastFaSyl(fName)
}
\arguments{
\item{fName}{The name of the netCDF file to be loaded}
}
\value{
The function returns a table with 2 columns, the first one with the years and the second one with
the associated probability of the mast event
}
\description{
Calculate the mast probability over a time series, including the current year,
using as input the file downloaded from the Copernicus CDS API.
}
\examples{
mastFaSyl("39434_t2p_tp.nc")

}
