% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastPlot.R
\name{mastPlot}
\alias{mastPlot}
\title{Plots the calculated mast probability of a site}
\usage{
mastPlot(predictionT)
}
\arguments{
\item{predictionT}{It is the table returned by the mastBeech function}
}
\value{
Returns the plot with a line chart of the predicted mast event probability
}
\description{
Plot the calculated mast probability on a line chart.
}
\examples{
 dataT <- mastFaSyl("inst/85930_t2s_tp.nc")
           plot <- mastPlot(dataT)

}
