% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probPlot.R
\name{probPlot}
\alias{probPlot}
\title{Plots the calculated mast probability of a site}
\usage{
probPlot(prediction)
}
\arguments{
\item{prediction}{It is the table returned by the mastFaSyl function}
}
\value{
Returns the plot with a line chart of the predicted mast event probability
}
\description{
Plot the calculated mast probability on a line chart.
}
\examples{
\dontrun{
data <- mastFaSyl("39434_t2p_tp.nc")
finalPlot <- probPlot(data)
}

}
