% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsDownload.R
\name{cdsDownload}
\alias{cdsDownload}
\title{Download of the climate data.}
\usage{
cdsDownload(U_ID, API_Key, lat, lon, sPath, site_id = "")
}
\arguments{
\item{U_ID}{The user ID for the authentication in the Copernicus CDS service (string)}

\item{API_Key}{The key needed to make an API request (string)}

\item{lat}{The latitude of the site of interest, expressed in decimals with a single decimal digit.
For southing latitude values use a negative number (integer)}

\item{lon}{The longitude of the site of interest, expressed in decimals with a single decimal digit.
For westing longitude values, use a negative number (integer)}

\item{sPath}{The path of the directory where the file is going to be saved (string)}

\item{site_id}{Set as empty, it is a string where it is possible to insert the study area id, which will be
reported in the file name}
}
\description{
Create a .nc file, containing the mean temperature and precipitation records of a defined location,
downloaded from the ECMFWR "ERA-5 Land monthly averaged data", via the Copernicus CDS API.
Before using, the registration to the Copernicus CDS site is required (https://cds.climate.copernicus.eu/#!/home)
as well as the acceptance of the terms and conditions here: https://cds.climate.copernicus.eu/cdsapp/#!/terms/licence-to-use-copernicus-products.
}
\examples{

}
