% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mast.R
\name{mastBeech}
\alias{mastBeech}
\title{Beech mast probability calculation}
\usage{
mastBeech(filename)
}
\arguments{
\item{filename}{The name of the .nc file to be loaded as input for the analysis}
}
\value{
The function returns a table with 2 columns, the first one with the years and the second one with
the associated probability of the mast event
}
\description{
Calculate the mast probability for the current year, using as input the file downloaded from the Copernicus
CDS API. It prints a plot representing a time series of the mast event probabilty from the first available
data to the current year
}
\examples{
mastProbability <- mastBeech(filename = "85930_t2s_tp.nc")
}
