% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastPlot.R
\name{mastPlot}
\alias{mastPlot}
\title{Plots the calculated mast probability of a site}
\usage{
mastPlot(predictionT, printPlot = FALSE)
}
\arguments{
\item{predictionT}{It is the table returned by the mastBeech function}

\item{printPlot}{To print the plot as pdf, set the parameter to TRUE (FALSE by default)}
}
\value{
If printPlot == FALSE, returns the plot without printing it
}
\description{
Plot the calculated mast probability on a line chart. The plot can be saved as pdf setting
printPlot = TRUE
}
\examples{
mastProbability <- mastBeech(filename = "85930_t2s_tp.nc")
          plot <- mastPlot(predictionT = mastProbability)
}
