% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastPlot.R
\name{mastPlot}
\alias{mastPlot}
\title{Plots the calculated mast probability of a site}
\usage{
mastPlot(prediction)
}
\arguments{
\item{prediction}{It is the table returned by the mastFaSyl function}
}
\value{
Returns the plot with a line chart of the predicted mast event probability
}
\description{
Plot the calculated mast probability on a line chart.
}
