% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastPlot.R
\name{mastPlot}
\alias{mastPlot}
\title{Plots the calculated mast probability of a site}
\usage{
mastPlot(predictionT, printPlot = FALSE)
}
\arguments{
\item{predictionT}{It is the table returned by the mastBeech function}
}
\value{

}
\description{
Save a
}
